Add North and East lanes calculation

+++ src/TrafficSimulator.Presentation.WPF/ViewModels/IntersectionElements/LaneElement.cs
@@ -5,8 +5,8 @@ namespace TrafficSimulator.Presentation.WPF.ViewModels.IntersectionElements
 	public class LaneElement
 	{
 		public int Width { get; set; }
-		public int StartPointX { get; set; }
-		public int StartPointY { get; set; }
+		public double StartPointX { get; set; }
+		public double StartPointY { get; set; }
 		public WorldDirection WorldDirection { get; set; }
 		public bool Inbound { get; set; }
 	}
